Move dicecoeff tests to bloommatcher tests; move common bitarray utilities to their own file.
